+++ src/main/java/com/rmnnorbert/InquireNet/dao/UserRowMapper.java
@@ -10,7 +10,7 @@ public class UserRowMapper implements RowMapper<User> {
     @Override
     public User mapRow(ResultSet rs, int rowNum) throws SQLException {
         return new User(
-                rs.getInt("id"),
+                rs.getLong("id"),
                 rs.getString("status"),
                 rs.getString("name"),
                 rs.getString("password"),
