+++ src/main/java/com/rmnnorbert/InquireNet/dto/answer/VoteDTO.java
@@ -1,4 +1,4 @@
 package com.rmnnorbert.InquireNet.dto.answer;
 
-public record VoteDTO (String vote, int id){
+public record VoteDTO (String vote, long id){
 }
